Add the ability to cluster nodes
The idea here is you can use present to connect to a cluster and make
variables available in the environment, which can then be written to
the config before starting the app. I think there would be two steps
where you'd do:

  we -s `present nodes $some_node_ip` \
     toconfig -t /etc/foo.conf.tmpl -s /etc/foo.conf \
     present join $some_node_ip \
     myapp

Eventually I'll make this all run via a single command with a config
file, but for now it'll look pretty verbose.